Update so.js
Changed script to execute on document start to reduce tearing, which meant no more jQuery, sad face, and also required attaching an observer temporarily (it is only around for roughly 40ms so don't worry about long term performance degradation). There is still the $ variable, but it is merely a facade for querySelector (much like jQuery is anyway).